+++ apps/main/src/llamalend/features/market-position-details/BorrowPositionDetails.stories.tsx
@@ -13,7 +13,8 @@ import type { Address } from '@primitives/address.utils'
 import type { Decimal } from '@primitives/decimal.utils'
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { getTokenUsdRateKey } from '@ui-kit/lib/model/entities/token-usd-rate'
-import { TestQueryProvider, useFakeMarket } from '@ui-kit/lib/queries/query.test-util'
+import { TestQueryProvider } from '@ui-kit/lib/queries/test-query.provider.test'
+import { useFakeMarket } from '@ui-kit/lib/queries/useFakeMarket.test'
 import type { Range } from '@ui-kit/types/util'
 import { CRVUSD_ADDRESS } from '@ui-kit/utils'
 import { BorrowPositionDetails } from './'
